+++ Changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+-   Changed babel plugin that made builds differ between windows and other
+    platforms.
+
 ## 6.1.0 - 2022-05-06
 
 ### Added